+++ sd/source/ui/func/fuprlout.cxx
@@ -149,7 +149,7 @@ void FuPresentationLayout::DoExecute( SfxRequest& rReq )
         SdAbstractDialogFactory* pFact = SdAbstractDialogFactory::Create();
         AbstractSdPresLayoutDlg* pDlg = pFact ? pFact->CreateSdPresLayoutDlg(mpDocSh, NULL, aSet ) : 0;
 
-        sal_uInt16 nResult = pDlg ? pDlg->Execute() : RET_CANCEL;
+        sal_uInt16 nResult = pDlg ? pDlg->Execute() : static_cast<short>(RET_CANCEL);
 
         switch (nResult)
         {
